Change approach
Prior we were generating the certs when the server was run but each time a save happened, it caused the certs to be regenerated. On Mac that mean supplying my password on every save which is no bueno. There may be a better way to handle this but for now, it will work and I won't go crazy in the meantime.